miniproject/requirements.txt
Flask
flask-mail
python-dotenv
mysql-connector-python
qrcode
Pillow
fpdf
groq
gunicorn
numpy
opencv-python-headless
face-recognition
